requirements.txt
# Core Framework
python-dotenv
pydantic
pydantic-settings

# Database
sqlalchemy
alembic
psycopg2-binary
asyncpg

# HTTP Client
httpx

# LangChain & LLM
langchain>=0.1.0
langchain-core
langchain-openai
langchain-anthropic
langchain-google-genai  # Gemini integration for LangChain
langgraph  # Modern agent framework for LangChain
openai
anthropic
google-generativeai  # Gemini API (Form Generator)

# Utilities
fuzzywuzzy
python-Levenshtein

# Streamlit UI
streamlit
plotly
pandas
nest-asyncio

# Document Processing
openpyxl
xlrd
python-docx
weasyprint
pypdf  # PDF form filling (Population Module)
pdfplumber  # PDF layout analysis (Form Generator)
reportlab  # PDF form field creation (Form Generator)
Pillow  # Image processing (Form Generator)
pdf2image  # PDF to image conversion (Form Generator)
jinja2

# Date/Time
python-dateutil

# Testing
pytest
pytest-asyncio
pytest-cov

# Development
black
mypy
ruff
